+++ bluetooth/base.py
@@ -1,12 +1,12 @@
-import serial
-import os
-# # Serial port parameters
-serial_speed = os.getenv("BlUETOOTH_SPEED")
-serial_port = os.getenv("BlUETOOTH_PORT") # bluetooth shield hc-06
+import logging
 
+# Enable logging
+logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
+                    level=logging.INFO)
+
+logger = logging.getLogger(__name__)
 
 class Base:
-    def __init__(self, msgid, name):
-        self.msgid = msgid
-        self.ser = serial.Serial(serial_port, serial_speed, timeout=1)
-        self.name = name
+    def __init__(self, msgid):
+        self.logger = logger.getChild(self.name)
+        self.id = msgid