+++ bluetooth/sender.py
@@ -1,15 +1,13 @@
-from .base import Base
+from .connection import Connection
 import time
-import logging
 
-# Enable logging
-logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-                    level=logging.INFO)
-
-logger = logging.getLogger(__name__)
-
-class Sender(Base):
+class Sender(Connection):
+    """
+        Class to send to arduino a msg
+        message_format: {id}_{msg}
+    """
 
     def send(self, msg):
-        logger.info('Sending msg')
-        self.ser.write(str.encode(msg))
+        encoded_msg = str(self.id) + "_" + msg
+        self.logging.info("Sending msg: " + encoded_msg)
+        self.ser.write(str.encode(encoded_msg))