+++ db/__init__.py
@@ -1,0 +1,4 @@
+from .setup import Setup
+from .schedule import Schedule
+
+__all__ = ('Setup', 'Schedule')