+++ jobs/__init__.py
@@ -1,0 +1,3 @@
+from .schedules import SchedulesJob
+
+__all__ = ('SchedulesJob')